+++ nextgisweb/webmap/view.py
@@ -199,12 +199,15 @@ def display(obj, request):
 
         elif item.item_type in ("root", "group"):
             expanded = item.group_expanded
+            exclusive = item.group_exclusive
             if expanded:
                 items_states.get("expanded").append(item.id)
             # Recursively run all elements excluding those
             # with no permissions
             data.update(
-                expanded=expanded, children=list(filter(None, map(traverse, item.children)))
+                expanded=expanded,
+                exclusive=exclusive,
+                children=list(filter(None, map(traverse, item.children))),
             )
             # Hide empty groups
             if (item.item_type in "group") and not data["children"]: